+++ weblate_schemas/test_valid.py
@@ -90,3 +90,33 @@ def test_userdata():
         },
         load_schema("weblate-userdata.schema.json"),
     )
+
+
+def test_backup():
+    """Test memory schema being valid."""
+    validate(
+        {
+            "metadata": {
+                "version": "4.13",
+                "server": "Weblate",
+                "domain": "weblate.example.com",
+                "timestamp": "2021-11-18T18:53:54.862Z",
+            },
+            "project": {
+                "name": "Hello",
+                "slug": "hello",
+                "web": "https://weblate.org/",
+                "instructions": "",
+                "set_language_team": False,
+                "use_shared_tm": False,
+                "contribute_shared_tm": False,
+                "access_control": 0,
+                "translation_review": False,
+                "source_review": False,
+                "enable_hooks": False,
+                "language_aliases": "",
+            },
+            "labels": [],
+        },
+        load_schema("weblate-backup.schema.json"),
+    )